rename project and source code folder from cookie_cutter_2 to modeling_learner_behavior
