--- cookie_cutter_2/__init__.py
@@ -1,1 +1,0 @@
-from cookie_cutter_2 import config  # noqa: F401
