+++ docs/mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: cookie_cutter_2
+site_name: modeling_learner_behavior
 # 
 site_author: Ren
 #
