+++ modeling_learner_behavior/__init__.py
@@ -1,0 +1,1 @@
+from modeling_learner_behavior import config  # noqa: F401
