+++ modeling_learner_behavior/dataset.py
@@ -1,0 +1,39 @@
+from pathlib import Path
+
+import typer
+from loguru import logger
+from tqdm import tqdm
+
+from modeling_learner_behavior.config import PROCESSED_DATA_DIR, RAW_DATA_DIR
+#modeling_learner_behavior.config import PROCESSED_DATA_DIR, RAW_DATA_DIR
+
+app = typer.Typer()
+
+
+@app.command()
+def main(
+    # ---- REPLACE DEFAULT PATHS AS APPROPRIATE ----
+    input_path: Path = RAW_DATA_DIR / "dataset.csv",
+    output_path: Path = PROCESSED_DATA_DIR / "dataset.csv",
+    # ----------------------------------------------
+):
+    # ---- REPLACE THIS WITH YOUR OWN CODE ----
+    logger.info("Processing dataset...")
+    for i in tqdm(range(10), total=10):
+        if i == 5:
+            logger.info("Something happened for iteration 5.")
+    logger.success("Processing dataset complete.")
+    # -----------------------------------------
+
+
+def do_something():
+    """this function prints hi
+
+    Returns:
+        _type_: _description_
+    """
+    return "hi"
+
+if __name__ == "__main__":
+    app()
+
