+++ modeling_learner_behavior/features.py
@@ -4,7 +4,7 @@ import typer
 from loguru import logger
 from tqdm import tqdm
 
-from cookie_cutter_2.config import PROCESSED_DATA_DIR, RAW_DATA_DIR
+from modeling_learner_behavior.config import PROCESSED_DATA_DIR
 
 app = typer.Typer()
 
@@ -12,16 +12,16 @@ app = typer.Typer()
 @app.command()
 def main(
     # ---- REPLACE DEFAULT PATHS AS APPROPRIATE ----
-    input_path: Path = RAW_DATA_DIR / "dataset.csv",
-    output_path: Path = PROCESSED_DATA_DIR / "dataset.csv",
-    # ----------------------------------------------
+    input_path: Path = PROCESSED_DATA_DIR / "dataset.csv",
+    output_path: Path = PROCESSED_DATA_DIR / "features.csv",
+    # -----------------------------------------
 ):
     # ---- REPLACE THIS WITH YOUR OWN CODE ----
-    logger.info("Processing dataset...")
+    logger.info("Generating features from dataset...")
     for i in tqdm(range(10), total=10):
         if i == 5:
             logger.info("Something happened for iteration 5.")
-    logger.success("Processing dataset complete.")
+    logger.success("Features generation complete.")
     # -----------------------------------------
