+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "cookie_cutter_2"
+name = "modeling_learner_behavior"
 version = "0.0.1"
 description = "another cookie cutter data science test"
 authors = [
@@ -28,5 +28,5 @@ exclude = '''
 '''
 
 [tool.ruff.lint.isort]
-known_first_party = ["cookie_cutter_2"]
+known_first_party = ["modeling_learner_behavior"]
 force_sort_within_sections = true
